requirements-test.txt
pytest==8.3.4
pytest-asyncio==0.20.3
pytest-homeassistant-custom-component==0.13.205
